The window width and height are now retrieved with love.window.getMode rather than love.graphics.getDimensions.
The coordinate space of the latter is always in pixels, even when the desktop's coordinate space isn't (e.g. in OS X on a Retina display with high-dpi mode enabled), but the window dragging calculations should always be in the desktop's coordinate space. love.window.getMode returns the width and height of the window in the desktop's coordinate space / scale.